Add BaseGame abstraction and refactor games
